directly export AccessToken, TokenCredential, TokenRequestOptions
