+++ sdk/azcore/internal/exported/exported.go
@@ -7,39 +7,13 @@
 package exported
 
 import (
-	"context"
 	"errors"
 	"io"
 	"io/ioutil"
 	"net/http"
-	"time"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/internal/shared"
 )
-
-// TokenRequestOptions contain specific parameter that may be used by credentials types when attempting to get a token.
-// Exported as policy.TokenRequestOptions.
-type TokenRequestOptions struct {
-	// Scopes contains the list of permission scopes required for the token.
-	Scopes []string
-	// TenantID contains the tenant ID to use in a multi-tenant authentication scenario, if TenantID is set
-	// it will override the tenant ID that was added at credential creation time.
-	TenantID string
-}
-
-// TokenCredential represents a credential capable of providing an OAuth token.
-// Exported as azcore.TokenCredential.
-type TokenCredential interface {
-	// GetToken requests an access token for the specified set of scopes.
-	GetToken(ctx context.Context, options TokenRequestOptions) (*AccessToken, error)
-}
-
-// AccessToken represents an Azure service bearer access token with expiry information.
-// Exported as azcore.AccessToken.
-type AccessToken struct {
-	Token     string
-	ExpiresOn time.Time
-}
 
 type nopCloser struct {
 	io.ReadSeeker
