+++ sdk/azcore/policy/policy.go
@@ -108,7 +108,13 @@ type TelemetryOptions struct {
 }
 
 // TokenRequestOptions contain specific parameter that may be used by credentials types when attempting to get a token.
-type TokenRequestOptions = exported.TokenRequestOptions
+type TokenRequestOptions struct {
+	// Scopes contains the list of permission scopes required for the token.
+	Scopes []string
+	// TenantID contains the tenant ID to use in a multi-tenant authentication scenario, if TenantID is set
+	// it will override the tenant ID that was added at credential creation time.
+	TenantID string
+}
 
 // BearerTokenOptions configures the bearer token policy's behavior.
 type BearerTokenOptions struct {
